communityID column type changed on int for App domain

+++ persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/App.java
@@ -29,7 +29,7 @@ public class App implements Serializable {
 	@Column(name="model",columnDefinition="char(40)")
 	private String model;
 	
-	private byte communityID;
+	private int communityID;
 
     public App() {
     }
